fix (dev_spring_boot) renamed user service classes
