+++ src/main/java/com/project/user/service/UserService.java
@@ -1,82 +1,17 @@
 package com.project.user.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.fge.jsonpatch.JsonPatchException;
 import com.github.fge.jsonpatch.mergepatch.JsonMergePatch;
 import com.project.order.model.OrderDTO;
-import com.project.order.service.DefaultOrderService;
-import com.project.user.model.Role;
-import com.project.user.model.User;
 import com.project.user.model.UserDTO;
-import com.project.user.repository.UserRepository;
-import com.project.utils.exceptionhandler.exceptions.InvalidUpdateRequest;
-import com.project.utils.exceptionhandler.exceptions.NoSuchElemException;
-import com.project.utils.mappers.EntityDtoMapper;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
 
-import java.text.MessageFormat;
+public interface UserService {
+    Page<UserDTO> getAll(Pageable pageable);
 
-import static com.project.utils.exceptionhandler.ExceptionMessages.*;
+    UserDTO get(Long id);
 
-@Service
-@RequiredArgsConstructor
-public class UserService implements CrudUserService {
-    private final UserRepository userRepository;
-    private final EntityDtoMapper entityDtoMapper;
-    private final DefaultOrderService orderService;
+    Page<OrderDTO> getOrders(Long id, Pageable pageable);
 
-    @Override
-    public Page<UserDTO> getAll(Pageable pageable) {
-        return userRepository.findAll(pageable).map(entityDtoMapper::toUserDTO);
-    }
-
-    @Override
-    public UserDTO get(Long id) {
-        User currentUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        if (currentUser.getRole().equals(Role.ADMIN) || id.equals(currentUser.getId())) {
-            return entityDtoMapper.toUserDTO(userRepository.findById(id).orElseThrow(() -> new NoSuchElemException(MessageFormat.format(USER_NOT_FOUND_ID, id))));
-        }
-
-        throw new AccessDeniedException(FORBIDDEN);
-    }
-
-    @Override
-    public Page<OrderDTO> getOrders(Long id, Pageable pageable) {
-        return orderService.getByUser(id, pageable);
-    }
-
-    @Override
-    public UserDTO update(Long id, JsonMergePatch patch) {
-        User attemptingUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        if (!id.equals(attemptingUser.getId())) {
-            throw new AccessDeniedException(FORBIDDEN);
-        }
-
-        User dbUser = userRepository
-                .findById(id)
-                .orElseThrow(() -> new NoSuchElemException(MessageFormat.format(USER_NOT_FOUND_ID, id)));
-
-        ObjectMapper objectMapper = new ObjectMapper();
-        UserDTO updatedUser;
-        try {
-            JsonNode updatedJson = patch.apply(objectMapper.convertValue(dbUser, JsonNode.class));
-            updatedUser = objectMapper.treeToValue(updatedJson, UserDTO.class);
-        } catch (JsonProcessingException | JsonPatchException e) {
-            throw new InvalidUpdateRequest(INVALID_USER_UPDATE);
-        }
-
-        dbUser.setName(updatedUser.getName());
-        dbUser.setSurname(updatedUser.getSurname());
-
-        return entityDtoMapper.toUserDTO(userRepository.save(dbUser));
-    }
+    UserDTO update(Long id, JsonMergePatch patch);
 }
